Moved marker along trajectory example to WellMarkersLayer story
