+++ typescript/packages/subsurface-viewer/src/storybook/layers/__image_snapshots__/subsurfaceviewer-well-markers-layer--markers-along-well-path.png
